LayoutTests/imported/w3c/web-platform-tests/focus/cross-origin-ancestor-activeelement-after-child-lose-focus.sub-expected.txt

CONSOLE MESSAGE: Blocked access to external URL http://www1.localhost:8800/focus/support/cross-origin-ancestor-activeelement-after-child-lose-focus-helper.html


Harness Error (TIMEOUT), message = null

PASS Ancestor's activeElement should be cleared when child loses focus
